Refactor[Auth]: 랭킹 조회 속도 빠르게 수정

+++ WebServer/src/main/java/ssafy/authserv/domain/record/dto/SpeedRankingInfo.java
@@ -7,12 +7,12 @@ import java.util.concurrent.atomic.AtomicLong;
 
 @Builder
 public record SpeedRankingInfo(
-        int rank,
+        long rank,
         String nickname,
         String profileImage,
         String time
 ) {
-    public static SpeedRankingInfo convertToDTO(Member member, String time, int ranking) {
+    public static SpeedRankingInfo convertToDTO(Member member, String time, long ranking) {
         return SpeedRankingInfo.builder()
                 .rank(ranking)
                 .nickname(member.getNickname())
